Combine_Neo4j_Yuran
Combine:
Neo4j, Knowledge Graph
Yuran

+++ AIops_BackEnd/src/main/java/com/aiops_web/entity/sql/ParsedLog.java
@@ -24,7 +24,7 @@ public class ParsedLog implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "parse_id", type = IdType.AUTO)
-    private Integer parseId;
+    private Long parseId;
 
     private String logLineid;
