+++ AIops_BackEnd/src/main/java/com/aiops_web/entity/sql/VectorizedLog.java
@@ -23,7 +23,7 @@ public class VectorizedLog implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "vector_id", type = IdType.AUTO)
-    private Integer vectorId;
+    private Long vectorId;
 
     private String embedding;
